Refactor SokosPostCommand, use SokosRequest and remove request encoding #deploy-skattekort-service
The request encoding and logging have been moved to SkattekortService. Now, SokosPostCommand uses SokosRequest instead of a String request. The SokosSkattekortConsumer has also been updated to accommodate the new SokosRequest parameter type. Meanwhile, we've changed the inntektsaar datatype in SkattekortTilArbeidsgiverDTO from Long to Integer.

+++ libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/skattekortservice/v1/SkattekortTilArbeidsgiverDTO.java
@@ -1,6 +1,5 @@
 package no.nav.testnav.libs.dto.skattekortservice.v1;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -56,7 +55,7 @@ public class SkattekortTilArbeidsgiverDTO {
         private Resultatstatus resultatPaaForespoersel;
         private Skattekort skattekort;
         private List<Tilleggsopplysning> tilleggsopplysning;
-        private Long inntektsaar;
+        private Integer inntektsaar;
 
         public List<Tilleggsopplysning> getTilleggsopplysning() {
